Added option parsing (main module)
Also added copyright, version, name to habitat/__init__.py, and made
sure that calling import on a top level module (utils, in this case)
causes sub modules to be loaded when they are meant to be used directly
(utils.dynamicloader. Note difference to message_server.server, which is
not used directly. message_server.server.Server is aliased to
message_server.Server)