Remove __all__ from modules, replacing with imports.
__all__ is only used for `from module import *` (outside of special uses cases
such as how loadable_manager inspects it; hence it stays for loadables such as
sensors and filters). As this style of import is strongly discouraged, remove
the __all__ statements (a `import *` will now import everything in the module
whose name does not start with an underscore).

In addition, "from . import submodule" is now in each of the `__init__.py`
files, so that executing `import habitat` allows use of `habitat.parser`
etcetera without explicit `import habitat.parser`.